workflow: use podman build with unconfined seccomp
The `chmod` in Fedora 41 uses `fchmodat2`. This is allowed in the
default seccomp from containers/common vendored in e.g. podman v5+, but
here we're running on ubuntu-latest which right now is 22.04, which has
an ancient podman (v3.4.4). Even 24.04 is still on v4.

Anyway, no point in trying to lock this down here, so just run
unconfined.